move the content process main to another file as well
